chore: rename root folder
