--- PancakeSwap-Bots/Cake-05/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-dist
-forta.config.json
