+++ PancakeSwap-Bots/Set-Masterchef-Settings/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "description": "Forta Agent Typescript starter project",
   "chainIds": [
-    97
+    56, 97
   ],
   "scripts": {
     "build": "tsc",
@@ -31,7 +31,6 @@
     "@types/nodemon": "^1.19.0",
     "jest": "^27.0.6",
     "nodemon": "^2.0.8",
-    "prettier": "^2.7.1",
     "ts-jest": "^27.0.3",
     "typescript": "^4.3.4"
   }
