--- PancakeSwap-Bots/Set-Masterchef-Settings/src/abi.ts
@@ -1,11 +1,0 @@
-const CAKE: string[] = [
-    "function setMigrator(IMigratorChef _migrator)",
-    "function dev(address _devaddr)",
-    "function add(uint256 _allocPoint, IBEP20 _lpToken, bool _withUpdate)",
-    "function set(uint256 _pid, uint256 _allocPoint, bool _withUpdate)",
-    "function updateMultiplier(uint256 multiplierNumber)",
-  ];
-  
-  export default {
-    CAKE,
-  };
