+++ PancakeSwap-Bots/Set-Masterchef-Settings/src/agent.ts
@@ -1,12 +1,38 @@
-import abi from "./abi";
-import findings from "./findings";
-import { TransactionEvent } from "forta-agent";
+import {
+  Finding,
+  HandleTransaction,
+  TransactionEvent,
+  getEthersProvider,
+} from "forta-agent";
+import { providers } from "ethers";
+import NetworkData from "./network";
+import NetworkManager from "./network";
+import utils from "./utils";
 
-const MasterChef_Address = "0xE115b296916E63C74E01506d37c6CF2C9fBaa9E0";
+const networkManager = new NetworkManager();
 
-export const provideHandleTransaction = (cakeAddress: string) => async (txEvent: TransactionEvent) =>
-  txEvent.filterLog(abi.CAKE, cakeAddress).map(findings.resolver);
+export const initialize = (provider: providers.Provider) => async () => {
+  const { chainId } = await provider.getNetwork();
+  networkManager.setNetwork(chainId);
+};
+
+export const handleTransaction =
+  (contractAddress: NetworkData): HandleTransaction =>
+  async (txEvent: TransactionEvent) => {
+    const findings: Finding[] = [];
+    const functionCalls = txEvent.filterFunction(
+      utils.FUNCTIONS_ABI,
+      contractAddress.factory
+    );
+
+    functionCalls.forEach((functionCall) => {
+      findings.push(utils.createFinding(functionCall, contractAddress));
+    });
+
+    return findings;
+  };
 
 export default {
-  handleTransaction: provideHandleTransaction(MasterChef_Address),
+  initialize: initialize(getEthersProvider()),
+  handleTransaction: handleTransaction(networkManager),
 };
